41: hard float Pt. 1

+++ toy/float.c
@@ -1,7 +1,7 @@
 // 2023-06-04 16:52
 #include <assert.h>
 
-float foo(float a, float b) { return a + b + (double)0.1; }
+float foo(float a, float b) { return a + b + 0.1f; }
 int main(int argc, char *argv[]) {
   float x = foo(0.1f, 0.2f);
   assert(x == 0.4f);
